Moved the integration test to proper file.
